fix: remove chmod when attempting to write to file
If a file doesn't give its owner write permission, it may be for a good reason.
Also, if the file isn't writable because the user isn't the ownerof the file,
the chmod will fail.
This was causing issues in the case where a user has permission to write to
a file because they belong to the group that owns the file.